refactor: blockもコンポーネントにしちゃう
